Moving things around
Made some folders
Also added in easier functionality for creating levels and enemies
