+++ Dog/Assembly-CSharp.csproj
@@ -234,14 +234,14 @@
      <Compile Include="Assets\GPGSIds.cs" />
      <Compile Include="Assets\Scripts\BattleSystem\ABtnHit.cs" />
      <Compile Include="Assets\Scripts\BattleSystem\ABtnTouch.cs" />
-     <Compile Include="Assets\Scripts\BattleSystem\austin&apos;s battle scripts\EnemyContainer.cs" />
-     <Compile Include="Assets\Scripts\BattleSystem\austin&apos;s battle scripts\enemyHealthBarController.cs" />
-     <Compile Include="Assets\Scripts\BattleSystem\austin&apos;s battle scripts\EnemyStatsConfig.cs" />
-     <Compile Include="Assets\Scripts\BattleSystem\austin&apos;s battle scripts\LevelConfig.cs" />
-     <Compile Include="Assets\Scripts\BattleSystem\austin&apos;s battle scripts\NewBehaviourScript.cs" />
      <Compile Include="Assets\Scripts\BattleSystem\BattleHandler.cs" />
      <Compile Include="Assets\Scripts\BattleSystem\Buttons\RightButton.cs" />
+     <Compile Include="Assets\Scripts\BattleSystem\Dog Scripts\DogContainer.cs" />
+     <Compile Include="Assets\Scripts\BattleSystem\Dog Scripts\DogStatsConfig.cs" />
+     <Compile Include="Assets\Scripts\BattleSystem\Enemy Scripts\EnemyContainer.cs" />
+     <Compile Include="Assets\Scripts\BattleSystem\Enemy Scripts\EnemyStatsConfig.cs" />
      <Compile Include="Assets\Scripts\BattleSystem\Enemy.cs" />
+     <Compile Include="Assets\Scripts\BattleSystem\Level Scripts\LevelConfig.cs" />
      <Compile Include="Assets\Scripts\BattleSystem\SpBtnTouch.cs" />
      <Compile Include="Assets\Scripts\BattleSystem\TriggerEvent.cs" />
      <Compile Include="Assets\Scripts\Credits.cs" />
