+++ Dog/Assets/Scripts/BattleSystem/Dog Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 03aa2670b08caa34a9aa24254038e67a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
