+++ Dog/Assets/Scripts/BattleSystem/Dog Scripts/DogStatsConfig.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 47f71583814745d4fb42eb71a99b160d
+guid: 4972d64167df211478cc16fc8543273f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
