--- Dog/Assets/Scripts/BattleSystem/austin's battle scripts/enemyHealthBarController.cs
@@ -1,20 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-
-public class enemyHealthBarController : MonoBehaviour
-{
-    public Slider bar;
-    public EnemyStatsConfig maxStuff;
-    void Start()
-    {
-
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-
-    }
-}
