+++ Dog/Assets/Scripts/Enemies/brownSquirrelStatsConfig.asset.meta
