+++ Dog/Assets/Scripts/Levels/myCoolLevel.asset
@@ -14,7 +14,5 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   enemies:
   - {fileID: 11400000, guid: 7ed5ec31e7bbb7d47910c4a0a07d3718, type: 2}
+  - {fileID: 11400000, guid: 2e4792c57ac39664497d7a2a1f534846, type: 2}
   background: {fileID: 21300000, guid: 08db75587297da24396de76d486db47b, type: 3}
-  damageTaken: 0
-  damageDealt: 0
-  dogDied: 0
